feat!: add `commit::virtual_merge_base()` to produce the single merge-base to use.
This allows more flexibility in conjunction with tree-merging, as
commits as input aren't required.

This is breaking as it changes the return value of `commit()`.
